Finished the J,H,K lightcurve-with-flags development. Works well.
The color-color and color-mag parts are more problematic; I'm getting errors,
and I don't really want to have to do the full code development it would take
to have equivalent functionality to the lightcurves - it would be very messy
and take a while, and there's no obvious way to visualize the 3 different
types of data flags in the color scatterplots anyways, not to mention the
trouble of coordinating the colormaps.
I think I might make an executive choice of accepting all "info" flags
and rejecting all "warning" flags for those.