1: Upgraded Stetson_chooser() to Stetson_machine();
2: commented out `chip` and `get_chip` for future removal from statcruncher().

Note: Stetson_machine() is a much more comprehensive function, and I
could see it being moved to stetson.py someday when it's more mature.
It might someday get replaced by an "ultimate Stetson" function that combines
nights with 1, 2, or 3 datapoints in the optimally described way... but that'll
be a ways down the road from here, I hope, because it doesn't sound like fun to
code up.