Multiple corrections and upgrades to lsp_power().
What used to be "try/except" statements are now carefully controlled
"if/else" statements, so that we don't lose tracking on unforeseen errors
in timing.py that turn out to be simple problems in the default keyword
argument values for `season`. Also, the default keyword argument value
for `season` is fixed. Two problems solved.