Fixed some issues with the last commit that were not,
for some annoying reason, dealt with when I used "git commit --amend". Only the commit message was altered in that one, despite promises from the internet that all changes I'd made since initially committing would be updated in the amended commit.

I dunno how to deal with this going forward, because not understanding this piece of functionality - how to undo and redo a commit - seems like a serious gap in my working knowledge of git.