Now the chitest.in and diagnostic.txt files are deleted after use.
Also, they are generated in-place to ensure portability. (This is actually the main reason I changed the code; the delete-after-use is just to avoid clutter, a problem that was previously solved by simply putting them in an obscure location. Obviously this is impossible when you want your code to run somewhere else, so deleting-after-use is a better solution.)

I've tested the code after editing -- we can still find periods when giving plot3.graded_phase no period argument, and we can still generate plot3.fx2_periodogram plots, so I strongly believe nothing broke in the process. (This was a fairly simple refactor, I think, but I did make sure I tested it a little.)