Compressed what gets printed in spreadsheet_write_efficient.
Someday I'll add in a verbose kwarg when I figure out how to make it play nice with the **kwargs that are already sitting in there.

Also I'll figure out how to print multiple things to a line and watch stdout update live rather than in a batch all at the end.