Added 'plot4', currently just an EXPERIMENT to see if we can separate these things out into a data Object and a plotter function.
I'll probably want to futher alter this so that the data object has METHODS rather than a bunch of predefined code that just runs in the __init__ step. Then you could even ASK for which data you want and how you want it to be sliced according to flags, colors, timestamps, etc!